Oct 19 updates
Went through and double checked data. I found a couple of critical errors during analyses that are now fixed.

Updated Analyses to include Spring 2017. Also edited fall 2017 analyses to look at diversity contrasts between mono, poly3, and poly5s together and independently. Because poly3s are not a mix of all possible poly5 combinations analyzing the poly3s compared to average of all monocultures is problematic as some (2 sites to be precise) of those monos are not included in the poly3 treatments. To get around this I subsetted the data to only include sites that were also in the poly3 for the poly3 only analyses.